agent/pyproject.toml
[project]
name = "broker-intellj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.21.0",
    "ipykernel>=7.2.0",
    "neo4j>=6.0.3",
    "python-dotenv>=1.0.0",
]
